--- pytorch_lightning/trainer/connectors/env_vars_connector.py
@@ -1,40 +1,0 @@
-# Copyright The PyTorch Lightning team.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from functools import wraps
-from typing import Callable
-
-from pytorch_lightning.utilities.argparse import get_init_arguments_and_types, parse_env_variables
-
-
-def _defaults_from_env_vars(fn: Callable) -> Callable:
-    """Decorator for :class:`~pytorch_lightning.trainer.trainer.Trainer` methods for which input arguments should
-    be moved automatically to the correct device."""
-
-    @wraps(fn)
-    def insert_env_defaults(self, *args, **kwargs):
-        cls = self.__class__  # get the class
-        if args:  # inace any args passed move them to kwargs
-            # parse only the argument names
-            cls_arg_names = [arg[0] for arg in get_init_arguments_and_types(cls)]
-            # convert args to kwargs
-            kwargs.update(dict(zip(cls_arg_names, args)))
-        env_variables = vars(parse_env_variables(cls))
-        # update the kwargs by env variables
-        kwargs = dict(list(env_variables.items()) + list(kwargs.items()))
-
-        # all args were already moved to kwargs
-        return fn(self, **kwargs)
-
-    return insert_env_defaults
